% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_masking_ex1_ggp2.R
\name{app12.2.2}
\alias{app12.2.2}
\title{Data-masking ggplot2 app 1}
\usage{
app12.2.2()
}
\value{
shiny app
}
\description{
From https://mastering-shiny.org/action-tidy.html#example-ggplot2
}
\examples{
# app12.2.2()
}
