% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_masking_ex3_dplyr.R
\name{app12.2.3}
\alias{app12.2.3}
\title{Data-masking dplyr app}
\usage{
app12.2.3()
}
\value{
shiny app
}
\description{
From https://mastering-shiny.org/action-tidy.html#example-dplyr
}
\examples{
# app12.2.3()
}
