% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_masking_user_data.R
\name{app12.2.4}
\alias{app12.2.4}
\title{Data-masking user data}
\usage{
app12.2.4()
}
\value{
shiny app
}
\description{
From https://mastering-shiny.org/action-tidy.html#user-data
}
\examples{
# app12.2.4()
}
