% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eval_across_ex2.R
\name{app12.3.2v2}
\alias{app12.3.2v2}
\title{Tidy-selection and data-masking (version 2)}
\usage{
app12.3.2v2()
}
\value{
shiny app
}
\description{
From: https://mastering-shiny.org/action-tidy.html#tidy-selection-and-data-masking
}
\examples{
# app12.3.2v2()
}
