% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{app (server)}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{shiny server inputs}

\item{output}{shiny server inputs}

\item{session}{shiny server inputs}
}
\value{
shiny server argument for \code{shinyApp()}
}
\description{
app (server)
}
