% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{bin_maker}
\alias{bin_maker}
\title{Binary column maker}
\usage{
bin_maker(bin_type, size, missing = FALSE)
}
\arguments{
\item{bin_type}{binary column type}

\item{size}{how many rows?}

\item{missing}{logical, include missing?}
}
\description{
Binary column maker
}
\examples{
bin_maker(type = "log", size = 2)
bin_maker(type = "log", size = 2, missing = TRUE)
bin_maker(type = "int", size = 10)
bin_maker(type = "int", size = 10, missing = TRUE)
bin_maker(type = "chr", size = 10)
bin_maker(type = "chr", size = 10, missing = TRUE)
bin_maker(type = "fct", size = 10)
bin_maker(type = "fct", size = 10, missing = TRUE)
bin_maker(type = "ord", size = 10)
bin_maker(type = "ord", size = 10, missing = TRUE)
}
