% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{cat_maker}
\alias{cat_maker}
\title{Categorical column maker}
\usage{
cat_maker(type, size, lvls, missing = FALSE)
}
\arguments{
\item{type}{column type ('chr', 'fct', or 'ord')}

\item{size}{how many rows?}

\item{lvls}{how many levels?}

\item{missing}{logical, include missing?}
}
\description{
Categorical column maker
}
\examples{
cat_maker(type = "chr", size = 10, lvls = 4, missing = TRUE)
cat_maker(type = "fct", size = 10, lvls = 4, missing = TRUE)
cat_maker(type = "ord", size = 6, lvls = 5, missing = FALSE)
}
\keyword{internal}
