% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{check_binary_vec}
\alias{check_binary_vec}
\title{Binary checks (helper)}
\usage{
check_binary_vec(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{type of column to return.}
}
\value{
TRUE/FALSE if binary
}
\description{
Binary checks (helper)
}
\examples{
require(palmerpenguins)
check_binary_vec(palmerpenguins::penguins$sex, type = "fct")
check_binary_vec(palmerpenguins::penguins$species, type = "fct")
}
