% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{check_facet_vec}
\alias{check_facet_vec}
\title{Facet checks (helper)}
\usage{
check_facet_vec(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{type of column to return.}
}
\value{
TRUE/FALSE if facet variable (< 5 levels)
}
\description{
Facet checks (helper)
}
\examples{
require(NHANES)
check_facet_vec(NHANES::NHANES$Education, type = "fct")
levels(NHANES::NHANES$Education)
check_facet_vec(NHANES::NHANES$MaritalStatus, type = "fct")
levels(NHANES::NHANES$MaritalStatus)
}
