% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{chr_maker}
\alias{chr_maker}
\title{Character column maker}
\usage{
chr_maker(size, lvls = size/2, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{lvls}{how many levels?}

\item{missing}{logical, include missing?}
}
\description{
Character column maker
}
\examples{
chr_maker(size = 2, lvls = 2, TRUE)
chr_maker(size = 2, lvls = 1, TRUE)
chr_maker(size = 3, lvls = 3, FALSE)
chr_maker(size = 3, lvls = 3, TRUE)
chr_maker(size = 10, lvls = 6, TRUE)
}
