% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{col_maker}
\alias{col_maker}
\title{Column maker}
\usage{
col_maker(col_types, size, missing, ...)
}
\arguments{
\item{col_types}{column type}

\item{size}{how many rows?}

\item{missing}{logical, include missing?}

\item{...}{additional arguments passed to \verb{_maker} functions}
}
\description{
Column maker
}
\examples{
col_maker(col_types = c("log", "log", "dbl", "dbl", "fct", "ord", "chr"),
  size = 6, missing = FALSE)
col_maker(col_types = c("log", "log", "dbl", "dbl", "fct", "ord", "chr"),
  size = 10, missing = TRUE, lvls = 4)
col_maker(col_types = c("log", "log", "dbl", "dbl", "fct", "ord", "chr"),
  size = 10, missing = TRUE, lvls = 2)
col_maker(col_types = c("log", "log", "dbl", "dbl", "fct", "ord", "chr"),
  size = 6, missing = FALSE)
}
