% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_col_types.R
\name{col_type_nms}
\alias{col_type_nms}
\title{Column type names}
\usage{
col_type_nms(data, type)
}
\arguments{
\item{data}{\code{data.frame}/\code{tibble}}

\item{type}{any \code{is.[type]} function (i.e., \code{is.numeric},
\code{is.character}, \code{is.logical}, etc.)}
}
\value{
data.frame/tibble filtered by function
}
\description{
Column type names
}
\examples{
col_type_nms(mtcars, "int")
}
