% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_skim.R
\name{custom_skim}
\alias{custom_skim}
\title{Custom \code{skimr::skim()} for numeric variables}
\usage{
custom_skim(x)
}
\description{
A custom \code{skimr::skim_with()} for numeric variables.
}
\examples{
require(dplyr)
custom_skim(dplyr::starwars)
}
