% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetApp.R
\name{datasetApp}
\alias{datasetApp}
\title{Dataset app}
\usage{
datasetApp(filter = NULL)
}
\value{
shiny application
}
\description{
This standalone app function is from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
\section{filter}{

This application includes an argument for filtering objects "\emph{limit the options to built-in datasets that are either data frames (filter = is.data.frame) or matrices (filter = is.matrix)}" displayed in the \code{selectInput()}.
}

\section{reactive values}{

The \code{renderPrint()} output contains the output from \code{shiny::reactiveValuesToList()}
and contains the\code{inputId}s for the modules in the application.
}

