% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{datasetInput}
\alias{datasetInput}
\title{Data input module}
\usage{
datasetInput(id, filter = NULL)
}
\arguments{
\item{id}{module id}

\item{filter}{type of object (i.e. \code{is.data.frame} or \code{is.matrix})}
}
\value{
list of data object names
}
\description{
This module function has been adapted from the \href{https://mastering-shiny.org/scaling-modules.html#getting-started-ui-input-server-output}{Getting started: UI input + server output} section of Mastering Shiny.
}
