% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{dbl_maker}
\alias{dbl_maker}
\title{Double column maker}
\usage{
dbl_maker(size, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{missing}{logical, include missing?}
}
\description{
Double column maker
}
\examples{
dbl_maker(2)
dbl_maker(2, TRUE)
dbl_maker(3)
dbl_maker(3, TRUE)
dbl_maker(10)
dbl_maker(10, TRUE)
}
