% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugApp-package.R
\docType{package}
\name{debugApp-package}
\alias{debugApp}
\alias{debugApp-package}
\title{debugApp: A Shiny App-Package with various debugging techniques}
\description{
A Shiny app-package developed to demonstrate debugging.
}
\keyword{internal}
