% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_nms_from_pkgs.R
\name{df_nms_from_pkgs}
\alias{df_nms_from_pkgs}
\title{Get names of data.frames from package}
\usage{
df_nms_from_pkgs(pkg)
}
\arguments{
\item{pkg}{}
}
\value{
named vector of data.frames in package
}
\description{
Get names of data.frames from package
}
\examples{
df_nms_from_pkgs(pkg = "base")
df_nms_from_pkgs(pkg = "datasets")
}
