% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{facet_maker}
\alias{facet_maker}
\title{Facet column maker}
\usage{
facet_maker(facet_type, size, lvls, missing = FALSE)
}
\arguments{
\item{facet_type}{facet column type}

\item{size}{how many rows?}

\item{lvls}{how many levels?}

\item{missing}{logical, include missing?}
}
\description{
Facet column maker
}
\examples{
facet_maker(type = "chr", size = 10, lvls = 4, missing = TRUE)
facet_maker(type = "fct", size = 10, lvls = 4, missing = TRUE)
facet_maker(type = "ord", size = 6, lvls = 5, missing = FALSE)
}
