% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{fct_maker}
\alias{fct_maker}
\title{Factor column maker}
\usage{
fct_maker(size, lvls = size/2, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{lvls}{how many levels?}

\item{missing}{logical, include missing?}
}
\description{
Factor column maker
}
\examples{
fct_maker(size = 5, lvls = 6, missing = TRUE)
fct_maker(size = 5, missing = TRUE)
fct_maker(size = 10, lvls = 9, missing = TRUE)
fct_maker(size = 10, lvls = 5, missing = FALSE)
fct_maker(size = 10, lvls = 5, missing = FALSE)
}
