% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_col_types.R
\name{filter_col_types}
\alias{filter_col_types}
\title{Filter columns by function}
\usage{
filter_col_types(data, filter)
}
\arguments{
\item{data}{\code{data.frame}/\code{tibble}}

\item{filter}{any \code{is.[type]} function (i.e., \code{is.numeric},
\code{is.character}, \code{is.logical}, etc.)}
}
\value{
data.frame/tibble filtered by function
}
\description{
Filter columns by function
}
\examples{
filter_col_types(movies)
}
