% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vars.R
\name{find_vars}
\alias{find_vars}
\title{Find Variable Names Based on a Filter Function}
\usage{
find_vars(data, filter)
}
\arguments{
\item{data}{A data frame or list on which the \code{filter} will be applied.}

\item{filter}{A function that is applied to each column of \code{data}.
It should return a logical vector of length 1.}
}
\value{
A character vector of column names for which the filter function
returns \code{TRUE}.
}
\description{
This function applies a filter function to each column of a data frame
and returns the names of columns for which the filter function returns \code{TRUE}.
}
\examples{
df <- data.frame(a = 1:10, b = letters[1:10], c = rnorm(10))
find_vars(df, is.numeric)  # Returns names of numeric columns

}
\keyword{internal}
