% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_with_dfs.R
\name{get_pkgs_with_dfs}
\alias{get_pkgs_with_dfs}
\title{Get packages with data.frames}
\usage{
get_pkgs_with_dfs()
}
\value{
named vector of packages with \code{data.frame} objects
}
\description{
Get packages with data.frames
}
\examples{
require(dplyr)
require(tidyr)
get_pkgs_with_dfs()
}
