% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{gg_color_scatter_facet}
\alias{gg_color_scatter_facet}
\title{Colored point plot (scatter-plot) with facets (with bugs!)}
\usage{
gg_color_scatter_facet(df, x_var, y_var, col_var = NULL, facet_var = NULL, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{col_var}{color variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{facet_var}{facet variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{...}{other arguments passed to (\code{ggplot2::facet_wrap(vars())})}
}
\value{
A \code{ggplot2} plot object
}
\description{
Colored point plot (scatter-plot) with facets (with bugs!)
}
