% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistApp.R
\name{ggHistApp}
\alias{ggHistApp}
\title{\code{ggHistApp}: A Standalone Shiny Application for Dynamic Histogram Visualization}
\usage{
ggHistApp()
}
\value{
This function does not return a value but launches a Shiny application
with specified UI and server logic, capable of interactive histogram generation.
}
\description{
\code{ggHistApp()} launches a Shiny application dedicated to generating dynamic
histograms based on user-selected variables from a given dataset. The application
offers a sidebar for dataset and variable selection, and a main panel displaying
the histogram and the current application values in a textual format.
}
\details{
The application UI is defined with a fluid layout containing a sidebar
for input controls and a main panel for output displays. The sidebar allows users
to upload a dataset and select a variable from that dataset for histogram
visualization. The \code{datasetInput()} and \code{selectVarInput()} functions
provide UI elements for dataset upload and variable selection, respectively.
}
\note{
\code{ggHistApp()} function assumes \code{datasetInput()}, \code{selectVarInput()},
and \code{datasetServer()} are loaded from \code{mstsap} which is a dependency of
\code{msst2ap}. These components are  critical for the app's functionality,
handling data upload, variable  selection,and histogram rendering.
}
\section{Server side}{

On the server side, \code{datasetServer} handles the uploaded dataset, while
\code{ggSelectVarServer()} manages the selection of a variable from this dataset to be
used in the histogram.
}

\section{App function}{

The \code{ggHistServer()} module is then invoked to generate  and render the
histogram based on the selected variable. Additionally, the  application
provides a textual output (\code{output$vals}) that prints current input
values, facilitating debugging and user interaction tracking.
}

\examples{
# Simply call the function to launch the app
if (interactive()) {
  ggHistApp()
}

}
