% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSelectVarServer.R
\name{ggSelectVarServer}
\alias{ggSelectVarServer}
\title{\code{ggSelectVarServer()}: Shiny Module Server Function for Selecting Variables}
\usage{
ggSelectVarServer(id, data, filter = is.numeric)
}
\arguments{
\item{id}{A unique identifier for the module's namespace.}

\item{data}{A reactive expression that returns a dataset. This dataset is
used to populate the choices in the select input based on the filter
criteria.}

\item{filter}{A function used to determine which variables in the dataset
should be made available for selection. By default, it filters for
numeric variables. This function should not be reactive.}
}
\value{
Returns a reactive expression that contains the selected
variable's values from the dataset. If no variable is selected or if
the selected variable is not found in the dataset, it returns \code{NULL}.
}
\description{
\code{ggSelectVarServer()} creates a server-side module for a Shiny application
that allows users to select a variable from a reactive dataset. The choices
for the variables are filtered based on a provided function. By default,
it filters for numeric variables. It observes changes to the dataset and
updates the variable choices dynamically.
}
\section{Updates to selectVarServer}{

This shiny application is a modified version of the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{\code{histogramApp()}} to
use \code{ggplot2} functions.  The \code{selectVarServer()} module server function has
been slightly altered to return a single column \code{data.frame} (as opposed to
the original vector):
}

