% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramApp.R
\name{histogramApp}
\alias{histogramApp}
\title{Histogram App}
\usage{
histogramApp()
}
\value{
shiny app
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{Case study: histogram} section of Mastering Shiny.
}
\section{Histogram App}{

The \code{histogramApp()} is built using the \code{datasetInput()}/\code{datasetServer()},
\code{selectVarInput()}/\code{selectVarServer()} and \code{histogramOutput()}/\code{histogramServer()}
modules.
}

