% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramOutput.R
\name{histogramOutput}
\alias{histogramOutput}
\title{Histogram UI module}
\usage{
histogramOutput(id)
}
\arguments{
\item{id}{shiny module ui id}
}
\value{
shiny ui module
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{Case study: histogram} section of Mastering Shiny.
}
