% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramServer.R
\name{histogramServer}
\alias{histogramServer}
\title{Histogram server module}
\usage{
histogramServer(id, x, title = reactive("Histogram"))
}
\arguments{
\item{id}{shiny module server id}
}
\value{
shiny server module
}
\description{
This module function comes from the \href{https://mastering-shiny.org/scaling-modules.html#case-study-histogram}{Case study: histogram} section of Mastering Shiny.
}
