% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{int_maker}
\alias{int_maker}
\title{Integer column maker}
\usage{
int_maker(size, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{missing}{logical, include missing?}
}
\description{
Integer column maker
}
\examples{
int_maker(2)
int_maker(2, TRUE)
int_maker(3)
int_maker(3, TRUE)
int_maker(5, TRUE)
int_maker(10, TRUE)
}
