% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pkg_inst.R
\name{is_pkg_inst}
\alias{is_pkg_inst}
\title{Check if a package is installed, if not install it}
\usage{
is_pkg_inst(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\description{
Check if \code{pkg} is installed with \code{rlang::is_installed()}. If not, package
is installed with \code{pak::pkg_install()}.
}
\examples{
p <- c("emoji", "palmerpenguins")
 if (any(p  \%in\% loadedNamespaces())) {
  remove.packages(p)
 }
is_pkg_inst(c("emoji", "palmerpenguins"))

}
