% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{log_maker}
\alias{log_maker}
\title{Logical column maker}
\usage{
log_maker(size, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{missing}{logical, include missing?}
}
\description{
Logical column maker
}
\examples{
log_maker(2)
log_maker(2, TRUE)
log_maker(3)
log_maker(3, TRUE)
}
