% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_funs.R
\name{make_bin_vec}
\alias{make_bin_vec}
\title{Make vector of binary columns by type (tidyverse)}
\usage{
make_bin_vec(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of binary columns
}
\description{
Make vector of binary columns by type (tidyverse)
}
\examples{
require(palmerpenguins)
require(dplyr)
make_binary_vec(
  df = dplyr::select(
    palmerpenguins::penguins,
    dplyr::where(is.factor)
  ),
  type = "fct"
)
}
