% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{make_binary_vec}
\alias{make_binary_vec}
\title{Create vector of binary columns by type (helper)}
\usage{
make_binary_vec(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of binary columns
}
\description{
Create vector of binary columns by type (helper)
}
\examples{
require(palmerpenguins)
require(dplyr)
make_binary_vec(
  df = dplyr::select(
    palmerpenguins::penguins,
    dplyr::where(is.factor)
  ),
  type = "fct"
)
}
