% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter.R
\name{make_x_y_color_title}
\alias{make_x_y_color_title}
\title{Make x, y, color plot title}
\usage{
make_x_y_color_title(x, y, color)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make x, y, color plot title
}
\examples{
diamonds <- ggplot2::diamonds
mini_dmnds <- diamonds[sample(nrow(diamonds), 10000), ]
gg_scatter(
  df = mini_dmnds,
  x_var = "carat",
  y_var = "price",
  alpha = 1 / 3,
  color = "#000000",
  size = 2
) + ggplot2::labs(title =
make_x_y_color_title(
  x = "carat",
  y = "price",
  color = "cut"
))
}
