% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cols.R
\name{mod_cols_server}
\alias{mod_cols_server}
\title{Column Selection Server Module}
\usage{
mod_cols_server(id, ds)
}
\arguments{
\item{id}{A unique identifier for the module's namespace, matching the \code{id}
used in \code{mod_cols_ui}.}

\item{ds}{A reactive expression returning the dataset to be used for updating
column choices.}
}
\value{
Does not return a value but updates the UI elements for column
selection dynamically.
}
\description{
This server module handles the logic for dynamically updating
the column selection choices in the UI based on the dataset provided. It
filters and updates the choices for X, Y, Color, and Facet columns based on
their appropriate data types.
}
\section{Details}{

The function filters columns based on their data types (numeric for X and Y,
binary for Color, and suitable for Faceting) using a helper function
\code{pull_cols}. It then observes these reactive values to update the respective
UI elements accordingly.
}

\seealso{
\code{\link[shiny]{moduleServer}}, \code{\link[shiny]{reactive}}, \code{\link[shiny]{observe}},
\code{\link[shiny]{updateSelectInput}}, \code{\link[shiny]{bindEvent}}
}
