% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cols.R
\name{mod_cols_ui}
\alias{mod_cols_ui}
\title{Column Selection UI Module}
\usage{
mod_cols_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module's namespace.}
}
\value{
Returns a UI element consisting of four dropdown menus for selecting
columns.
}
\description{
This function creates a UI module for selecting columns for X,
Y, Color, and Facet variables in a dataset. It generates four dropdown menus
for each selection, initially filled with no choices.
}
\section{Details}{

The function provides UI elements for selecting columns that will later be
used in plotting or data analysis tasks. It is meant to be used in conjunction
with \code{mod_cols_server} to dynamically update the choices based on the dataset
provided to the server function.
}

\seealso{
\code{\link[shiny]{selectInput}}, \code{\link[shiny]{NS}}, \code{\link[shiny]{tagList}}
}
