% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_server}
\alias{mod_data_server}
\title{Data Selection Server Module}
\usage{
mod_data_server(id)
}
\arguments{
\item{id}{A unique identifier for the module's namespace, matching the \code{id} used in
\code{mod_data_ui}.}
}
\value{
Returns a reactive expression containing the selected dataset.
}
\description{
This function creates a server module to handle the selection of datasets
made through the UI module created by \code{mod_data_ui}. It outputs a reactive expression
that returns the dataset selected by the user.
}
\section{Details}{

The function listens to the input from the UI module and returns the corresponding dataset
as a reactive expression. It supports selecting the 'penguins' dataset from the
\code{palmerpenguins} package or the 'starwars' dataset from the \code{dplyr} package.
}

\seealso{
\code{\link[shiny]{moduleServer}}, \code{\link[shiny]{reactive}},
\code{\link[palmerpenguins]{penguins}}, \code{\link[dplyr]{starwars}}
}
