% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_ui}
\alias{mod_data_ui}
\title{Data Selection UI Module}
\usage{
mod_data_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module's namespace.}
}
\value{
Returns a UI element that is a dropdown menu for dataset selection.
}
\description{
This function creates a UI module for selecting datasets. It provides
a dropdown menu allowing the user to select between the 'penguins' dataset from
the \code{palmerpenguins} package and the 'starwars' dataset from the \code{dplyr} package.
}
\section{Details}{

This function utilizes \code{shiny::selectInput} to create the dropdown menu. It is designed
to be used in conjunction with the \code{mod_data_server} function to provide the functionality
of selecting and displaying the dataset in a Shiny application.
}

\seealso{
\code{\link[shiny]{selectInput}}, \code{\link[palmerpenguins]{penguins}},
\code{\link[dplyr]{starwars}}
}
