% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_input.R
\name{mod_ds_input_server}
\alias{mod_ds_input_server}
\title{Dataset input Server}
\usage{
mod_ds_input_server(id, pkg)
}
\arguments{
\item{id}{namespaced id}
}
\value{
module server function
}
\description{
Dataset input Server
}
\keyword{internal}
