% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_input.R
\name{mod_ds_input_ui}
\alias{mod_ds_input_ui}
\title{Dataset input UI}
\usage{
mod_ds_input_ui(id, pkg, filter = NULL)
}
\arguments{
\item{id}{namespaced id}

\item{pkg}{name of package}

\item{filter}{filter data by ...}
}
\value{
options for choices in \code{selectInput()}
}
\description{
Dataset input UI
}
\keyword{internal}
