% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds.R
\name{mod_ds_server}
\alias{mod_ds_server}
\title{Dataset server module}
\usage{
mod_ds_server(id, pkg_input)
}
\arguments{
\item{id}{module id}

\item{pkg_input}{selected package}
}
\value{
shiny server module
}
\description{
Dataset server module
}
\section{Returned object (\code{pkg_input})}{


The returned value passed to \code{pkg_input} is the result of the
\code{get_pkgs_with_dfs()}, which returns a character vector of all packages with
a \code{data.frame} or \code{tibble} in their exports.

\code{mod_pkg_server} is the precursor to \code{mod_ds_server}, and is intended
to be return a character vector of packages (see example below):

\if{html}{\out{<div class="sourceCode">}}\preformatted{pkg <- mod_pkg_server("pkg")

mod_ds_server("ds", pkg_input = pkg)
}\if{html}{\out{</div>}}
}

