% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_server}
\alias{mod_graph_server}
\title{Plot server module}
\usage{
mod_graph_server(id, graph_data)
}
\arguments{
\item{id}{module id}

\item{graph_data}{a \code{data.frame} from \code{mod_vars}}
}
\value{
shiny server module
}
\description{
Plot server module
}
