% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg.R
\name{mod_pkg_server}
\alias{mod_pkg_server}
\title{Package server module}
\usage{
mod_pkg_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny server module
}
\description{
Package server module
}
\section{Using \code{mod_pkg_server}}{


\code{mod_pkg} is the intended to be used with \code{mod_ds}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{pkg <- mod_pkg_server("pkg")

mod_ds_server("ds", pkg_input = pkg)
}\if{html}{\out{</div>}}
}

