% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_data.R
\name{mod_select_data_server}
\alias{mod_select_data_server}
\title{Select data module UI}
\usage{
mod_select_data_server(id, pkg, filter = is.numeric)
}
\arguments{
\item{id}{namespaced id}

\item{filter}{function passed to variable from dataset}
}
\value{
UI module
}
\description{
Select data module UI
}
\keyword{internal}
