% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{mod_select_data_ui}
\alias{mod_select_data_ui}
\title{Select data module UI}
\usage{
mod_select_data_ui(id, pkg)
}
\arguments{
\item{id}{namespaced id}
}
\value{
UI module
}
\description{
Select data module UI
}
