% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_var.R
\name{mod_select_var_server}
\alias{mod_select_var_server}
\title{Select variable module server}
\usage{
mod_select_var_server(id, data, filter = is.numeric)
}
\arguments{
\item{id}{namespaced id}

\item{data}{selected dataset}

\item{filter}{function for filtering data object}
}
\description{
Select variable module server
}
\keyword{internal}
