% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var.R
\name{mod_select_var_ui}
\alias{mod_select_var_ui}
\title{Select variable input UI}
\usage{
mod_select_var_ui(id)
}
\arguments{
\item{id}{namespaced id}
}
\value{
module UI function
}
\description{
Select variable input UI
}
