% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vars.R
\name{mod_vars_server}
\alias{mod_vars_server}
\title{Select variables Server module}
\usage{
mod_vars_server(id, ds_input)
}
\arguments{
\item{id}{module id}

\item{ds_input}{pkg data from}
}
\value{
shiny server module
}
\description{
Select variables Server module
}
\section{Returned object (\code{graph_values})}{


The returned object passed to \code{ds_input} is the result of:

\if{html}{\out{<div class="sourceCode">}}\preformatted{get(x = input$dataset, pos = paste0("package:", pkg_input()))
}\if{html}{\out{</div>}}
\itemize{
\item The \code{ds_input} is converted to \code{pkg_data()} and used to populate the
\code{selectInput()}s for the \code{x}, \code{y}, \code{color}, and \code{facet} inputs.
}
}

