% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vars.R
\name{mod_vars_ui}
\alias{mod_vars_ui}
\title{Select variables UI module}
\usage{
mod_vars_ui(id)
}
\arguments{
\item{id}{module id}

\item{pkg_data}{data for application}
}
\value{
shiny UI module
}
\description{
Select variables UI module
}
