% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msst2ap-package.R
\docType{package}
\name{msst2ap-package}
\alias{msst2ap}
\alias{msst2ap-package}
\title{msst2ap: Mastering Shiny Shiny Test 2 App Package}
\description{
This Shiny app-package demonstrates testing with testthat and shinytest2 using the examples from from Mastering Shiny (mastering-shiny.org).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/msst2ap/}
}

}
\keyword{internal}
