% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstsap-package.R
\docType{package}
\name{mstsap-package}
\alias{mstsap}
\alias{mstsap-package}
\title{mstsap: Mastering Shiny Test Server App Package}
\description{
This Shiny app-package demonstrates testing with testthat using the examples from from Mastering Shiny (mastering-shiny.org).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/mstsap/}
}

}
\keyword{internal}
