% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtthat.R
\name{ord_maker}
\alias{ord_maker}
\title{Ordered factor column maker}
\usage{
ord_maker(size, lvls = size/2, missing = FALSE)
}
\arguments{
\item{size}{how many rows?}

\item{lvls}{how many levels?}

\item{missing}{logical, include missing?}
}
\description{
Ordered factor column maker
}
\examples{
ord_maker(size = 5, lvls = 6, missing = TRUE)
ord_maker(size = 10, lvls = 5, missing = FALSE)
ord_maker(size = 10, missing = FALSE)
ord_maker(size = 10, lvls = 5, missing = TRUE)
ord_maker(size = 10, lvls = 5, missing = FALSE)
}
