% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_nms_with_dfs.R
\name{pkg_nms_with_dfs}
\alias{pkg_nms_with_dfs}
\title{Get packages with data.frames}
\usage{
pkg_nms_with_dfs()
}
\value{
named vector of packages with \code{data.frame} objects
}
\description{
Get packages with data.frames
}
\examples{
require(dplyr)
require(tidyr)
pkg_nms_with_dfs()
}
