% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{pull_binary_cols}
\alias{pull_binary_cols}
\title{Binary app inputs}
\usage{
pull_binary_cols(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a named character vector of binary column names
}
\description{
Binary app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
pull_binary_cols(palmerpenguins::penguins)
pull_binary_cols(dplyr::starwars)
}
