% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{pull_facet_cols}
\alias{pull_facet_cols}
\title{Facet app inputs}
\usage{
pull_facet_cols(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a named character vector or column from with < 5 unique levels
}
\description{
Facet app inputs
}
\section{Variables to use for facets}{


This function is designed to quickly determine which variables have an
appropriate number of categorical levels for using \code{ggplot2::facet_wrap()}
or \code{ggplot2::facet_grid()}
}

\examples{
require(dplyr)
require(NHANES)
pull_facet_cols(df = dplyr::starwars)
pull_facet_cols(df = NHANES::NHANES)

str(dplyr::select(
  dplyr::starwars,
  dplyr::all_of(pull_facet_cols(df = dplyr::starwars))
))
str(dplyr::select(
  NHANES::NHANES,
  dplyr::all_of(pull_facet_cols(df = NHANES::NHANES))
))
}
