% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{pull_numeric_cols}
\alias{pull_numeric_cols}
\title{Numeric app inputs}
\usage{
pull_numeric_cols(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a named character vector of integer and double column names
}
\description{
Numeric app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
pull_numeric_cols(palmerpenguins::penguins)
pull_numeric_cols(dplyr::starwars)
}
