% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugApp-package.R
\docType{package}
\name{sapkgs-package}
\alias{sapkgs}
\alias{sapkgs-package}
\title{sapkgs: Shiny Application-Packages (repo)}
\description{
Shiny App-Packages with various debugging, testing, and package development techniques
}
\keyword{internal}
