% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{select_column_class}
\alias{select_column_class}
\title{Select column class (helper)}
\usage{
select_column_class(df, class)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{class}{class of column to return.
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"list"}: list
}}
}
\value{
\code{tibble} of columns matching \code{class} (empty \code{tibble} if no columns
of specified type
exist)
}
\description{
Return column in data by \code{class()}
}
\examples{
require(dplyr)
select_column_class(df = dplyr::starwars, class = "chr")
select_column_class(df = dplyr::starwars, class = c("chr", "list"))
select_column_class(df = mtcars, class = "chr")
select_column_class(df = mtcars, class = c("chr", "list"))
}
