% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_server.R
\name{skim_server}
\alias{skim_server}
\title{Skim app (server)}
\usage{
skim_server(input, output, session)
}
\arguments{
\item{input}{shiny server inputs}

\item{output}{shiny server inputs}

\item{session}{shiny server inputs}
}
\value{
shiny server argument for \code{shinyApp()}
}
\description{
Skim app (server)
}
