% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimVarApp.R
\name{skimVarApp}
\alias{skimVarApp}
\title{Skim data variable app}
\usage{
skimVarApp(filter = is.numeric)
}
\description{
Select a dataset and variable from the package and view
a \code{skimr::skim()} output.
}
