% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st2ap-package.R
\docType{package}
\name{st2ap-package}
\alias{st2ap}
\alias{st2ap-package}
\title{st2ap: Build Your Shiny App Like a Package}
\description{
This shiny app was developed as a package using devtools and usethis.
}
\keyword{internal}
