% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_cmt}
\alias{test_cmt}
\title{Test comment helper}
\usage{
test_cmt(test, msg)
}
\arguments{
\item{test}{test name}

\item{msg}{test message}
}
\description{
Test comment helper
}
\examples{
test_cmt("some test", "test message")
}
