% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsap-package.R
\docType{package}
\name{tsap-package}
\alias{tsap}
\alias{tsap-package}
\title{tsap: Test Server App Package}
\description{
A Shiny app-package developed to demonstrate testing with testthat and Shiny's testServer() function.
}
\keyword{internal}
