% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utap-package.R
\docType{package}
\name{utap-package}
\alias{utap}
\alias{utap-package}
\title{utap: Unit Test Your Shiny App Package (utap)}
\description{
A series of utility functions for a shiny app developed as an R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjfrigaard/sapkgs/tree/utap}
}

}
\keyword{internal}
