% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_to_csl}
\alias{vec_to_csl}
\title{Create a comma-separated list from a vector}
\usage{
vec_to_csl(x)
}
\arguments{
\item{x}{vector}
}
\value{
comma, separated, & list
}
\description{
Create a comma-separated list from a vector
}
\examples{
vec_to_csl(x = mtcars$disp)
}
