% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vars.R
\name{find_vars}
\alias{find_vars}
\title{Find variables}
\usage{
find_vars(data, filter)
}
\arguments{
\item{data}{selected dataset}

\item{filter}{function for filtering data object}
}
\value{
filtered data
}
\description{
Find variables
}
