% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cols.R
\name{mod_cols_server}
\alias{mod_cols_server}
\title{Select variables Server module}
\usage{
mod_cols_server(id, ds_input)
}
\arguments{
\item{id}{module id}

\item{ds_input}{pkg data from}
}
\value{
shiny server module
}
\description{
Select variables Server module
}
\section{Returned object (\code{ds_input})}{


The returned object passed to \code{ds_input} is the result of:

\if{html}{\out{<div class="sourceCode">}}\preformatted{get(x = input$dataset, pos = paste0("package:", pkg_input()))
}\if{html}{\out{</div>}}
\itemize{
\item Where \code{pkg_input()} is the return value from \code{mod_pkg_server()}, and
\code{input$dataset} is the reactive value from \code{mod_ds_server()}
}
}

