% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vars.R
\name{pull_cat_cols}
\alias{pull_cat_cols}
\title{Categorical app inputs}
\usage{
pull_cat_cols(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a named character vector of character and factor column names
}
\description{
Categorical app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
pull_cat_cols(palmerpenguins::penguins)
pull_cat_cols(dplyr::starwars)
}
