% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cols.R
\name{pull_cols}
\alias{pull_cols}
\title{Pull names of column by type}
\usage{
pull_cols(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{one of \code{bin}, \code{cat}, \code{num}, or \code{facet}}
}
\value{
a named character vector of column names by type
}
\description{
Pull names of column by type
}
\examples{
require(palmerpenguins)
pull_cols(palmerpenguins::penguins, "bin")
pull_cols(palmerpenguins::penguins, "cat")
pull_cols(palmerpenguins::penguins, "facet")
}
