% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_class.R
\name{select_class}
\alias{select_class}
\title{Select Column Class}
\usage{
select_class(df, class, return_tbl = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} from which columns will be selected.}

\item{class}{Character vector specifying the class(es) of columns to select.
Supported values are:
\itemize{
\item "logical" ("lo")
\item "integer" ("in")
\item "double" ("do")
\item "character" ("ch")
\item "factor" ("fa")
\item "ordered" ("or")
\item "list" ("li")
}}

\item{return_tbl}{Logical indicating whether to return the result as a
\code{data.frame}. If \code{FALSE}, a vector of selected column names is returned.}
}
\value{
A \code{data.frame} or vector of column names, depending on \code{return_tbl}.
}
\description{
\code{select_class()} selects columns from a data.frame based on the specified
\code{class}. Options include logical, integer, double, character, factor, ordered,
and list column types.
}
\examples{
data(iris)
head(select_class(iris, "numeric"))
head(select_class(iris, "factor"))
select_class(iris, c("factor", "numeric"), FALSE)

}
