% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utap-package.R
\docType{package}
\name{utap-package}
\alias{utap}
\alias{utap-package}
\title{utap: Unit Test Your Shiny App Package}
\description{
This shiny app was developed as a package using devtools and usethis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/utap/}
}

}
\keyword{internal}
